One test does not pass but its mostly a file path thing
okay as far as i can tell, Array and Panel are ready to rumble. Array can create panels and store them. It can also write a json file using its panel data as well as read a json and make panels from it. It only has one function to compare panels to each other. That is a part that will evolve more and more though. As it is now I think it could be fed data and analyze it.

I will build more funcitonaltiy for comparing panels next.