Skspatial spikes the memory like crazy. About to eliminate it.
Sksspatial cause several GBs of memory spike. I am using it to get the plane of best fit equation which allows us to get a normal vector to the plane.

However the Ransac3D libray gives me the same thing with a difference that is on the order of 10^-4. Ransac3D does not spike the memory so I think thats the way to go.